cern

+++ src/pages/profile/ProfileMembersDrawersContent.tsx
@@ -4,29 +4,27 @@ import TextField from '@mui/material/TextField'
 import { deleteUser, User } from 'firebase/auth'
 import { deleteDoc, doc } from 'firebase/firestore'
 import { useEffect, useState } from 'react'
-import { useNavigate } from 'react-router-dom'
-// import supabase from 'src/baseApi/base'
+import { useLocation, useNavigate } from 'react-router-dom'
 import { dbservice } from 'src/baseApi/serverbase'
-import { useSelectors } from 'src/hooks'
-import { useTexts } from 'src/hooks'
+import useSelectors from 'src/hooks/useSelectors'
+import useTexts from 'src/hooks/useTexts'
 
-interface Props {
-  user: object
-}
-const ProfileMembersDrawersContent = ({ user }: Props) => {
+const ProfileMembersDrawersContent = () => {
   const [confirmEmail, setConfirmEmail] = useState(false)
   const [process, setProcess] = useState(false)
   const {
     mail,
     deleteAccount,
     noProcessingCard,
-    ongoingCards,
+    processingCard,
     toDeleteAccountInputMail,
     canDeleteAccountWhenYouHaveNoProcessingBorrwingOrLendingCard,
   } = useTexts()
   const navigate = useNavigate()
   const profile = useSelectors((state) => state.profile.value)
-
+  const {state} = useLocation()
+  const user = state?.element || profile
+  console.log(profile)
   const onChange = (event) => {
     const {
       target: { value },
@@ -38,28 +36,26 @@ const ProfileMembersDrawersContent = ({ user }: Props) => {
     }
   }
   const delist = async () => {
-    await deleteDoc(doc(dbservice, `members/${profile?.uid}`))
-    deleteUser(user)
-      .then(() => {
-        console.log(user)
-        // User deleted.
-      })
-      .catch((error) => {
-        console.log(error)
-        // An error ocurred
-        // ...
-      })
-    // const { data, error } = await supabase.auth.admin.deleteUser(user)
-    // if (data) {
-    //   console.log(data)
-    // } else {
-    //   console.log(error)
-    // }
-    navigate('/')
+    if (process && confirmEmail) {
+      await deleteDoc(doc(dbservice, `members/${profile?.uid}`))
+      deleteUser(user)
+        .then(() => {
+          console.log(user)
+          // User deleted.
+        })
+        .catch((error) => {
+          console.log(error)
+          // An error ocurred
+          // ...
+        })
+      navigate('/')
+    } else {
+      alert('Cannot delete your account. Please fulfill the conditions.')
+    }
   }
   useEffect(() => {
-    const createdCards = user.userData?.createdCards
-    const connectedCards = user.userData?.connectedCards
+    const createdCards = profile?.createdCards
+    const connectedCards = profile?.connectedCards
     const createdNumber = createdCards?.length || 0
     const connectedNumber = connectedCards?.length || 0
     if (createdNumber === 0 && connectedNumber === 0) {
@@ -73,37 +69,28 @@ const ProfileMembersDrawersContent = ({ user }: Props) => {
     <>
       <div className="p-5">
         {canDeleteAccountWhenYouHaveNoProcessingBorrwingOrLendingCard}
-        {process ? (
-          <div className="flex justify-center">
-            <Chip
-              label={noProcessingCard}
-              sx={{ bgcolor: '#7fc4bc', color: 'white' }}
-            />
-          </div>
-        ) : (
-          <div className="flex justify-center">
-            <Chip
-              label={ongoingCards}
-              sx={{
-                bgcolor: '#e76e50',
-                color: 'white',
-              }}
-            />
-          </div>
-        )}
-        <div>{toDeleteAccountInputMail}</div>
+        <div className="flex justify-center">
+          <Chip
+            label={process ? 
+              noProcessingCard
+               : 
+              <>
+                Borrowing on process: {profile?.createdCards.length}
+                <br />
+                Lending on process: {profile?.connectedCards.length}
+              </>
+              //  processingCard
+              }
+            sx={{ bgcolor: process ? '#7fc4bc' : '#e76e50', color: 'white', paddingY: process ? 0 : 3 }}
+          />
+        </div>
+        <div className='pt-5'>{toDeleteAccountInputMail}</div>
       </div>
-      <div className="flex flex-col justify-center p-5 gap-5">
+      <div className="flex flex-col justify-center px-5 gap-5">
         <TextField label={mail} onChange={onChange} />
-        {process && confirmEmail ? (
-          <Button variant="outlined" onClick={delist}>
-            {deleteAccount}
-          </Button>
-        ) : (
-          <Button variant="outlined" disabled>
-            {deleteAccount}
-          </Button>
-        )}
+        <Button variant="outlined" onClick={delist}>
+          {deleteAccount}
+        </Button>
       </div>
     </>
   )
